pyproject.toml
[tool.poetry]
name = "wechat-guardian"
version = "0.1.0"
description = "A WeChat window guardian application"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
pystray = "^0.19.5"
Pillow = "^10.2.0"
pyautogui = "^0.9.54"
psutil = "^5.9.8"
pywin32 = "^308"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
wechat-guardian = "src.main:main"
